TK-01151 - Fix IMUSANT_duration::asAbsoluteNumeric() to handle cumulative floating point errors. Fixed by returning a float instead of a long.  We'll just have to put up with floats in the consuming algorithms ;-(.

+++ imusant old/IMUSANT/DomainObjects/inc/IMUSANT_duration.h
@@ -77,7 +77,7 @@ namespace IMUSANT
         }
         
         IMUSANT_duration	getSimplifiedDuration() const;
-        long                asAbsoluteNumeric() const;
+        float                asAbsoluteNumeric() const;
         
         //checks that TRational type is actually dotted note, return dots and changes dur
         static long	NormaliseDuration(TRational& dur);
